move EthService/wallets to WalletService
