+++ common/v2/services/EthService/index.ts
@@ -2,18 +2,6 @@ export { getNonce } from './nonce';
 export { Contract, ERC20, encodeTransfer, decodeTransfer } from './contracts';
 export { Web3Node, isWeb3Node, Web3Service, setupWeb3Node, RPCRequests, RPCNode } from './nodes';
 export {
-  determineKeystoreType,
-  KeystoreTypes,
-  signWrapper,
-  getKeystoreWallet,
-  getUtcWallet,
-  IWallet,
-  getPrivKeyWallet,
-  IFullWallet,
-  Web3Wallet,
-  MnemonicWallet
-} from './wallet';
-export {
   makeExplorer,
   stripHexPrefix,
   stripHexPrefixAndLower,
