--- common/v2/services/EthService/wallet/index.ts
@@ -1,4 +1,0 @@
-export { IWallet, IReadOnlyWallet, IFullWallet } from './IWallet';
-export { Balance } from './balance';
-export * from './deterministic';
-export * from './non-deterministic';
