+++ common/v2/services/WalletService/index.ts
@@ -2,3 +2,19 @@ export * from './mnemonic';
 export * from './privatekey';
 export * from './web3';
 export * from './keystore';
+export { IWallet, IReadOnlyWallet, IFullWallet } from './IWallet';
+export { Balance } from './balance';
+export * from './deterministic';
+export * from './non-deterministic';
+// export {
+//   determineKeystoreType,
+//   KeystoreTypes,
+//   signWrapper,
+//   getKeystoreWallet,
+//   getUtcWallet,
+//   IWallet,
+//   getPrivKeyWallet,
+//   IFullWallet,
+//   Web3Wallet,
+//   MnemonicWallet
+// } from './wallet';
